student_management/requirements.txt
asgiref==3.8.1
Django==3.2
djangorestframework==3.12.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
